--- NotebookAnalyzers/config.py
@@ -1,3 +1,0 @@
-from pathlib import Path
-
-data_path = Path("../data/").expanduser()
