--- NotebookAnalyzers/linters.py
@@ -1,8 +1,0 @@
-class NbLinter:
-    id = "nb-linter"
-    description = "Lints Python Jupyter notebooks, one at a time."
-
-
-class RepoLinter:
-    id = "repo-linter"
-    description = "Lints all Jupyter notebooks in a Python project."
