+++ pynblint/__main__.py
@@ -1,0 +1,10 @@
+""" Entry point of pynblint. Used when running pynblint from the command line. """
+
+
+def main():
+    # TODO: develop the command line interface of pynblint
+    pass
+
+
+if __name__ == '__main__':
+    main()
