+++ tests/unit/test_pynblint.py
@@ -1,0 +1,6 @@
+from pynblint.notebook import Notebook
+from pathlib import Path
+
+if __name__ == '__main__':
+    nb = Notebook(Path('../fixtures/my-attempt-at-analytics-vidhya-job-a-thon.ipynb'))
+    print(nb.get_pynblint_results())
